+++ core/core.php
@@ -1393,6 +1393,18 @@ class common {
 			$this->setData(['config', 'connect','captcha', true]);
 			$this->setData(['core', 'dataVersion', 10302]);
 		}
+		// Version 10.3.03
+		if ($this->getData(['core', 'dataVersion']) < 10303) {
+			// Activation par défaut du captcha à la connexion
+			$this->setData(['config', 'captchaStrong', false]);
+			$this->setData(['core', 'dataVersion', 10303]);
+		}
+		// Version 10.3.04
+		if ($this->getData(['core', 'dataVersion']) < 10304) {
+			// Couleur des sous menus
+			$this->setData(['theme', 'menu', 'backgroundColorSub', $this->getData(['theme', 'menu', 'backgroundColor']) ]);
+			$this->setData(['core', 'dataVersion', 10304]);
+		}
 		// Version 10.4.00
 		if ($this->getData(['core', 'dataVersion']) < 10400) {
 			// Ajouter le prénom comme pseudo et le pseudo comme signature
@@ -1437,12 +1449,6 @@ class common {
 				}
 			}
 			$this->setData(['core', 'dataVersion', 10400]);
-		}
-		// Version 10.3.03
-		if ($this->getData(['core', 'dataVersion']) < 10303) {
-			// Activation par défaut du captcha à la connexion
-			$this->setData(['config', 'captchaStrong', false]);
-		$this->setData(['core', 'dataVersion', 10303]);
 		}
 	}
 }
@@ -1590,7 +1596,7 @@ class core extends common {
 			$css .= 'header span{color:' . $colors['normal'] . ';font-family:"' . str_replace('+', ' ', $this->getData(['theme', 'header', 'font'])) . '",sans-serif;font-weight:' . $this->getData(['theme', 'header', 'fontWeight']) . ';font-size:' . $this->getData(['theme', 'header', 'fontSize']) . ';text-transform:' . $this->getData(['theme', 'header', 'textTransform']) . '}';
 			// Menu
 			$colors = helper::colorVariants($this->getData(['theme', 'menu', 'backgroundColor']));
-			$css .= 'nav,nav a{background-color:' . $colors['normal'] . '}';
+			$css .= 'nav,nav.navMain a{background-color:' . $colors['normal'] . '}';
 			$css .= 'nav a,#toggle span,nav a:hover{color:' . $this->getData(['theme', 'menu', 'textColor']) . '}';
 			$css .= 'nav a:hover{background-color:' . $colors['darken'] . '}';
 			$css .= 'nav a.active{color:' . $this->getData(['theme','menu','activeTextColor']) . ';}';
@@ -1602,8 +1608,11 @@ class core extends common {
 				$css .= 'nav a.active{color:' .  $color2['text'] . '}';*/
 			}
 			$css .= 'nav #burgerText{color:' .  $colors['text'] . '}';
-			$css .= 'nav .navLevel1 a.active {border-radius:' . $this->getData(['theme', 'menu', 'radius']) . '}';
-			$css .= '#menuLeft{float:' . $this->getData(['theme', 'menu', 'textAlign']) . '}';
+			// Sous menu
+			$colors = helper::colorVariants($this->getData(['theme', 'menu', 'backgroundColorSub']));
+			$css .= 'nav .navSub a{background-color:' . $colors['normal'] . '}';
+			$css .= 'nav .navMain a.active {border-radius:' . $this->getData(['theme', 'menu', 'radius']) . '}';
+			$css .= '#menu{text-align:' . $this->getData(['theme', 'menu', 'textAlign']) . '}';
 			if($this->getData(['theme', 'menu', 'margin'])) {
 				if(
 					$this->getData(['theme', 'menu', 'position']) === 'site-first'
@@ -1723,7 +1732,9 @@ class core extends common {
 			exit();
 		}
 		// Journalisation
-		$dataLog = strftime('%d/%m/%y',time()) . ';' . strftime('%R',time()) . ';' ;
+		$dataLog = mb_detect_encoding(strftime('%d/%m/%y',time()), 'UTF-8', true)
+					? strftime('%d/%m/%y',time()) . ';' . strftime('%R',time()) . ';'
+					: utf8_encode(strftime('%d/%m/%y',time())) . ';' . utf8_encode(strftime('%R',time())) . ';' ;
 		$dataLog .= helper::getIp() . ';';
 		$dataLog .= $this->getUser('id') ? $this->getUser('id') . ';' : 'anonyme' . ';';
 		$dataLog .= $this->getUrl();
@@ -2413,7 +2424,7 @@ class layout extends common {
 				empty($childrenPageIds)) {
 				continue;
 			}
-			$itemsLeft .= '<ul class="navLevel2">';
+			$itemsLeft .= '<ul class="navSub">';
 			foreach($childrenPageIds as $childKey) {
 				// Propriétés de l'item
 				$active = ($childKey === $currentPageId) ? 'active ' : '';
@@ -2487,7 +2498,7 @@ class layout extends common {
 			$itemsRight .= '<li><a id="barLogout" href="' . helper::baseUrl() . 'user/logout" data-tippy-content="Me déconnecter">' . template::ico('logout') . '</a></li>';
 		}
 		// Retourne les items du menu
-		echo '<ul class="navLevel1" id="menuLeft">' . $itemsLeft . '</ul><ul class="navLevel1" id="menuRight">' . $itemsRight . '</ul>';
+		echo '<ul class="navMain" id="menuLeft">' . $itemsLeft . '</ul><ul class="navMain" id="menuRight">' . $itemsRight . '</ul>';
 	}
 
 	/**
@@ -2525,7 +2536,7 @@ class layout extends common {
 				continue;
 			}
 			// Propriétés de l'item
-			$active = ($parentPageId === $currentPageId OR in_array($currentPageId, $childrenPageIds)) ? 'active ' : '';
+			$active = ($parentPageId === $currentPageId OR in_array($currentPageId, $childrenPageIds)) ? ' class="active"' : '';
 			$targetBlank = $this->getData(['page', $parentPageId, 'targetBlank']) ? ' target="_blank"' : '';
 			// Mise en page de l'item;
 			// Ne pas afficher le parent d'une sous-page quand l'option est sélectionnée.
@@ -2535,7 +2546,7 @@ class layout extends common {
 					AND $this->getUser('password') !== $this->getInput('ZWII_USER_PASSWORD')	) {
 						$items .= '<a href="'.$this->getUrl(1).'">';
 				} else {
-						$items .= '<a href="' . $active . helper::baseUrl() . $parentPageId . '"' . $targetBlank . '>';
+						$items .= '<a href="'. helper::baseUrl() . $parentPageId . '"' . $targetBlank . $active .'>';
 				}
 				$items .= $this->getData(['page', $parentPageId, 'title']);
 				$items .= '</a>';
@@ -2548,7 +2559,7 @@ class layout extends common {
 				}
 
 				// Propriétés de l'item
-				$active = ($childKey === $currentPageId) ? 'active ' : '';
+				$active = ($childKey === $currentPageId) ? ' class="active"' : '';
 				$targetBlank = $this->getData(['page', $childKey, 'targetBlank']) ? ' target="_blank"' : '';
 				// Mise en page du sous-item
 				$itemsChildren .= '<li class="menuSideChild">';
@@ -2557,7 +2568,7 @@ class layout extends common {
 					AND $this->getUser('password') !== $this->getInput('ZWII_USER_PASSWORD')	) {
 						$itemsChildren .= '<a href="'.$this->getUrl(1).'">';
 				} else {
-					$itemsChildren .= '<a href="' .$active . helper::baseUrl() . $childKey . '"' . $targetBlank . '>';
+					$itemsChildren .= '<a href="' . helper::baseUrl() . $childKey . '"' . $targetBlank . $active . '>';
 				}
 
 				$itemsChildren .= $this->getData(['page', $childKey, 'title']);