+++ module/blog/view/article/article.php
@@ -3,8 +3,14 @@
 		<div class="col10">
 			<div class="blogDate">
 				<i class="far fa-calendar-alt"></i>
-				<?php echo strftime('%d %B %Y', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn'])); ?>
-					à <?php echo strftime('%H:%M', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn'])); ?>
+				<?php $date = mb_detect_encoding(strftime('%d %B %Y', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn'])), 'UTF-8', true)
+								? strftime('%d %B %Y', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn']))
+								: utf8_encode(strftime('%d %B %Y', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn'])));
+					  $heure =  mb_detect_encoding(strftime('%H:%M', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn'])), 'UTF-8', true)
+					  			? strftime('%H:%M', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn']))
+								:  utf8_encode(strftime('%H:%M', $this->getData(['module', $this->getUrl(0), $this->getUrl(1), 'publishedOn'])));
+					  echo $date . ' à ' . $heure; 
+				?>
 			</div>
 		</div>
 			<div class="col2">
@@ -126,8 +132,10 @@
 			<?php foreach($module::$comments as $commentId => $comment): ?>
 				<div class="block">
 					<h4><?php echo $module::$commentsSignature[$commentId]; ?>
-						le <?php echo strftime('%d %B %Y - %H:%M', $comment['createdOn']); ?>
-					</h4>
+					le <?php  echo mb_detect_encoding(strftime('%d %B %Y - %H:%M', $comment['createdOn']), 'UTF-8', true)
+										 ? strftime('%d %B %Y - %H:%M', $comment['createdOn'])
+										 : utf8_encode(strftime('%d %B %Y - %H:%M', $comment['createdOn']));
+					?>
 					<?php echo $comment['content']; ?>
 				</div>
 			<?php endforeach; ?>